requirements.txt
pinecone-client==2.1.0
sentence-transformers==2.2.2
streamlit==1.17.0
pandas==1.5.2